+++ hari_client/client/client.py
@@ -2628,6 +2628,47 @@ class HARIClient:
             success_response_item_model=list[models.AttributeMetadataResponse],
         )
 
+    def delete_attribute_metadata(
+        self, dataset_id: uuid.UUID, attribute_metadata_id: str
+    ) -> None:
+        """Delete an attribute metadata and all its attributes from a dataset.
+
+        Args:
+            dataset_id: The ID of the dataset.
+            attribute_metadata_id: The ID of the attribute metadata
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "DELETE",
+            f"/datasets/{dataset_id}/attributeMetadata/{attribute_metadata_id}",
+            success_response_item_model=str,
+        )
+
+    def update_attribute_metadata(
+        self,
+        dataset_id: uuid.UUID,
+        attribute_metadata_id: str,
+        name: str | None = None,
+    ) -> models.AttributeMetadataResponse:
+        """Update attribute metadata.
+
+        Args:
+            dataset_id: The ID of the dataset.
+            attribute_metadata_id: The ID of the attribute metadata
+            name: The desired name of the attribute metadata
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "PATCH",
+            f"/datasets/{dataset_id}/attributeMetadata/{attribute_metadata_id}",
+            json=name,
+            success_response_item_model=models.AttributeMetadataResponse,
+        )
+
     def get_visualisation_configs(
         self,
         dataset_id: uuid.UUID,
@@ -3740,3 +3781,110 @@ class HARIClient:
             params=self._pack(locals()),
             success_response_item_model=models.AnnotationRunProjectDetails,
         )
+
+    ### annotations ###
+    def get_annotation(
+        self,
+        dataset_id: uuid.UUID,
+        annotation_id: str,
+    ) -> models.AnnotationResponse:
+        """Returns an annotation for the given annotation_id.
+
+        Args:
+            dataset_id: The dataset id
+            annotation_id: The annotation id
+
+        Returns:
+            The annotation with the given annotation_id
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            f"/datasets/{dataset_id}/annotations/{annotation_id}",
+            success_response_item_model=models.AnnotationResponse,
+        )
+
+    def get_annotations(
+        self,
+        dataset_id: uuid.UUID,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+    ) -> list[models.AnnotationResponse]:
+        """Get all annotations of a dataset
+
+        Args:
+            dataset_id: The dataset id
+            limit: The number of annotations tu return
+            skip: The number of annotations to skip
+            query: The filters to be applied to the search
+            sort: The list of sorting parameters
+
+        Returns:
+            A list of all annotations in a dataset
+
+        Raises:
+            APIException: If the request fails.
+        """
+        if limit is None:
+            return self.get_annotations_paginated(
+                dataset_id=dataset_id,
+                query=query,
+                sort=sort,
+            )
+
+        return self._request(
+            "GET",
+            f"/datasets/{dataset_id}/annotations",
+            params=self._pack(locals(), ignore=["dataset_id"]),
+            success_response_item_model=list[models.AnnotationResponse],
+        )
+
+    def get_annotations_paginated(
+        self,
+        dataset_id: uuid.UUID,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        batch_size: int = 100,
+    ) -> list[models.AnnotationResponse]:
+        """Get all annotations of a dataset but with pagination, could be used for larger amount of annotations.
+
+        Args:
+            dataset_id: The dataset id
+            query: The filters to be applied to the search
+            sort: The list of sorting parameters
+            batch_size: The number of annotations to fetch per request. Defaults to 100.
+
+        Returns:
+            A list of all annotations in a dataset
+
+        Raises:
+            APIException: If the request fails.
+        """
+
+        log.info(f"Fetching all annotations of the dataset {dataset_id} ...")
+
+        annotations: list[models.DatasetResponse] = []
+        skip_offset = 0
+
+        while True:
+            annotations_page = self.get_annotations(
+                dataset_id=dataset_id,
+                limit=batch_size,
+                skip=skip_offset,
+                query=query,
+                sort=sort,
+            )
+            annotations.extend(annotations_page)
+
+            if len(annotations_page) < batch_size:
+                break
+
+            skip_offset += batch_size
+
+        log.info(f"Fetched {len(annotations)} annotations successfully.")
+
+        return annotations