+++ scripts/data_upload/upload_dataset_with_own_annotation_attributes.py
@@ -200,7 +200,10 @@ def upload_dataset_with_own_attributes(
             media_object.set_object_category_subset_name(majority_vote)
 
     check_and_upload_dataset(
-        hari, dataset_id, data.categories, medias=list(medias.values())
+        hari=hari,
+        dataset_id=dataset_id,
+        object_categories=data.categories,
+        medias=list(medias.values()),
     )